feat: add markdown renderer and add basic styling
