+++ frontend/package.json
@@ -24,6 +24,7 @@
     "prisma": "^4.6.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-markdown": "^8.0.3",
     "wagmi": "^0.7.13"
   },
   "devDependencies": {
