+++ frontend/styles/globals.css
@@ -18,4 +18,3 @@ a {
 * {
   box-sizing: border-box;
 }
-
